Rearrange routes and server setup
also use `fastify-plugin` for the routes
