+++ packages/server/src/main.ts
@@ -3,6 +3,6 @@ import { server } from './server';
 server.listen({ port: 8888 }, (error) => {
   if (error) {
     server.log.error(error);
-    process.exit(1);
+    server.close().finally(() => process.exit(1));
   }
 });
