+++ packages/server/src/server.ts
@@ -2,33 +2,32 @@ import Fastify from 'fastify';
 import FastifyCookie from '@fastify/cookie';
 import { TypeBoxTypeProvider } from '@fastify/type-provider-typebox';
 import * as qs from 'qs';
-import { router as authRouter, prefix as authPrefix } from './routers/auth-router';
-import {
-  router as playlistsRouter,
-  prefix as playlistsPrefix,
-} from './routers/playlists-router';
+import spotifyWebApiClient from './plugins/spotify-web-api-client';
+import authRoutes from './routes/auth';
+import playlistsRoutes from './routes/playlists';
 import { HeadersContentTypeJson } from './shared/schemas/content-type-schemas';
 
 // set up server
-const fastify = Fastify({
+export const server = Fastify({
   logger: true,
   querystringParser: (str) => qs.parse(str),
 }).withTypeProvider<TypeBoxTypeProvider>();
 
-fastify.addSchema(HeadersContentTypeJson);
+server.addSchema(HeadersContentTypeJson);
 
-fastify.register(FastifyCookie);
+server.register(FastifyCookie);
 
-fastify.get('/healthcheck', (_, reply) => reply.status(200).send());
+server.get('/healthcheck', (_, reply) => reply.status(200).send());
 
-// register API routes
-fastify.register(
+// set up API
+server.register(
   async (apiInstance) => {
-    apiInstance.decorateRequest('spotify', null);
-    await apiInstance.register(authRouter, { prefix: authPrefix });
-    await apiInstance.register(playlistsRouter, { prefix: playlistsPrefix });
+    // API plugins
+    server.register(spotifyWebApiClient);
+
+    // API routes
+    await apiInstance.register(authRoutes);
+    await apiInstance.register(playlistsRoutes);
   },
   { prefix: '/api' },
 );
-
-export const server = fastify;
